feat: Complete quiz services domain migration (8 services → DDD)
✅ Migration Complete - Quiz Services to Domain

Migrated 8 quiz-related services from /app/services to Domain-Driven Design
architecture in /app/domain with zero breaking changes.

## Services Migrated

1. QuizTemplateService → domain/quizzes/templates/
2. QuizMetricsCollector → domain/analytics/quiz/
3. QuizLinkResilienceService → domain/quizzes/resilience/
4. QuizResponseEvaluator → domain/quizzes/evaluation/
5. Response utilities → domain/quizzes/utils/
6. Token rotation → domain/quizzes/security/
7. Flow integration (2 files) → domain/quizzes/integration/

## New Domain Structure

app/domain/
├── quizzes/ (6 subdomains: templates, evaluation, resilience, security, utils, integration)
└── analytics/quiz/ (metrics collection)

## Backward Compatibility

✅ All old import paths still work via deprecation adapters
✅ DeprecationWarnings guide developers to new paths
✅ Zero test failures
✅ Zero API changes

## Architecture Benefits

- Clear separation of concerns
- Better discoverability
- Easier testing and mocking
- Follows DDD principles
- Modular and scalable

## Domain Completion Status

- Quizzes: 95% complete (19 files)
- Analytics/Quiz: 100% complete (2 files)

## References

- Full documentation: docs/migrations/QUIZ_SERVICES_MIGRATION.md
- Original location: /app/services/quiz_*.py
- New location: /app/domain/quizzes/* and /app/domain/analytics/quiz/*
