+++ backend-hormonia/.claude-flow/metrics/performance.json
@@ -1,7 +1,7 @@
 {
-  "startTime": 1762542263452,
-  "sessionId": "session-1762542263452",
-  "lastActivity": 1762542263452,
+  "startTime": 1762544900453,
+  "sessionId": "session-1762544900453",
+  "lastActivity": 1762544900453,
   "sessionDuration": 0,
   "totalTasks": 1,
   "successfulTasks": 1,
