+++ backend-hormonia/.claude-flow/metrics/task-metrics.json
@@ -1,10 +1,10 @@
 [
   {
-    "id": "cmd-hooks-1762542264288",
+    "id": "cmd-hooks-1762544901332",
     "type": "hooks",
     "success": true,
-    "duration": 23.762024999999994,
-    "timestamp": 1762542264312,
+    "duration": 26.588882999999896,
+    "timestamp": 1762544901359,
     "metadata": {}
   }
 ]
