+++ backend-hormonia/app/domain/analytics/quiz/__init__.py
@@ -1,0 +1,4 @@
+"""Quiz analytics and metrics domain."""
+from .metrics_collector import QuizMetricsCollector, get_quiz_metrics_collector
+
+__all__ = ["QuizMetricsCollector", "get_quiz_metrics_collector"]
