+++ backend-hormonia/app/domain/quizzes/integration/__init__.py
@@ -1,0 +1,3 @@
+"""Quiz integration with flows and other systems."""
+# Import integration services here when they exist
+__all__ = []
