+++ backend-hormonia/app/domain/quizzes/resilience/__init__.py
@@ -1,0 +1,14 @@
+"""Quiz link resilience and fallback management."""
+from .link_resilience import (
+    QuizLinkResilienceService,
+    FailureReason,
+    CircuitBreakerState,
+    ResilienceMetrics
+)
+
+__all__ = [
+    "QuizLinkResilienceService",
+    "FailureReason",
+    "CircuitBreakerState",
+    "ResilienceMetrics"
+]
