+++ backend-hormonia/app/domain/quizzes/security/__init__.py
@@ -1,0 +1,10 @@
+"""Quiz security features including token rotation."""
+from .token_rotation import (
+    _validate_token_with_grace_period,
+    submit_quiz_response_with_rotation
+)
+
+__all__ = [
+    "_validate_token_with_grace_period",
+    "submit_quiz_response_with_rotation"
+]
