+++ backend-hormonia/app/domain/quizzes/utils/__init__.py
@@ -1,0 +1,16 @@
+"""Quiz utility functions."""
+from .response_utils import (
+    normalize_other_value,
+    serialize_response_value,
+    deserialize_response_value,
+    validate_multi_select_response,
+    extract_other_text_requirement
+)
+
+__all__ = [
+    "normalize_other_value",
+    "serialize_response_value",
+    "deserialize_response_value",
+    "validate_multi_select_response",
+    "extract_other_text_requirement"
+]
